Move legacy backbone classes to trackster and toolshed/reports grid wrapper
